refactor landing
