+++ app/javascript/gobierto_data/webapp/pages/Visualizations.vue
@@ -1,40 +1,62 @@
 <template>
   <div class="gobierto-data gobierto-data-landing-visualizations">
     <div class="pure-g gutters m_b_1">
-      <div class="pure-u-1 pure-u-lg-4-4 gobierto-data-layout-column">
-        <h3 class="gobierto-data-index-title">
-          {{ labelVisualizations }}
-        </h3>
-        <VisualizationsAllList
-          v-if="datasetsArray.length && datasetsVisualizations.length"
-          :datasets-array="datasetsArray"
-          :datasets-attributes="datasetsVisualizations"
+      <template v-if="!isVizsLoaded">
+        <SkeletonSpinner
+          height-square="250px"
+          square-rows="2"
+          squares="2"
         />
-      </div>
+      </template>
+      <template v-else>
+        <div class="pure-u-1 pure-u-lg-4-4 gobierto-data-layout-column">
+          <h3 class="gobierto-data-index-title">
+            {{ labelVisualizations }}
+          </h3>
+          <VisualizationsGrid
+            v-if="showVizs"
+            :public-visualizations="publicVisualizations"
+          />
+        </div>
+      </template>
     </div>
   </div>
 </template>
 <script>
-import VisualizationsAllList from "./../components/landingviz/VisualizationsAllList";
+import VisualizationsGrid from "./../components/landingviz/VisualizationsGrid";
 import { CategoriesMixin } from "./../../lib/mixins/categories.mixin";
+import { convertToCSV } from "./../../lib/helpers";
 import { FiltersMixin } from "./../../lib/mixins/filters.mixin";
 import { VisualizationFactoryMixin } from "./../../lib/factories/visualizations";
+import { DataFactoryMixin } from "./../../lib/factories/data";
+import { QueriesFactoryMixin } from "./../../lib/factories/queries";
+import { SkeletonSpinner } from "lib/vue-components";
 
 export default {
   name: "Visualizations",
   components: {
-    VisualizationsAllList
+    VisualizationsGrid,
+    SkeletonSpinner
   },
   mixins: [
     CategoriesMixin,
     FiltersMixin,
-    VisualizationFactoryMixin
+    VisualizationFactoryMixin,
+    DataFactoryMixin,
+    QueriesFactoryMixin
   ],
   data() {
     return {
       datasetsArray: [],
       datasetsVisualizations: [],
+      publicVisualizations: [],
+      isVizsLoaded: false,
       labelVisualizations: I18n.t("gobierto_data.projects.visualizations") || ""
+    }
+  },
+  computed: {
+    showVizs() {
+      return this.datasetsArray.length && this.datasetsVisualizations.length
     }
   },
   created() {
@@ -65,6 +87,72 @@ export default {
         }
       }
       this.datasetsVisualizations = this.subsetItems.filter((dataset) => datasetsWithVizs.includes(dataset.id))
+      this.getPublicVisualizations()
+    },
+    async getPublicVisualizations() {
+      let allVizs = []
+      for (let index = 0; index < this.datasetsArray.length; index++) {
+        const { data: response } = await this.getVisualizations({
+          "filter[dataset_id]": this.datasetsArray[index]
+        });
+        const { data } = response;
+
+        if (data.length) {
+          this.publicVisualizations = await this.getDataFromVisualizations(data);
+          allVizs.push(this.publicVisualizations);
+        }
+      }
+
+      this.publicVisualizations = allVizs.reduce((flatten, arr) => [...flatten, ...arr])
+      this.isVizsLoaded = true
+      this.removeAllIcons()
+    },
+    async getDataFromVisualizations(data) {
+      const visualizations = [];
+      for (let index = 0; index < data.length; index++) {
+        const { attributes = {}, id } = data[index];
+        const { query_id, user_id, sql = "", spec = {}, name = "", privacy_status = "open", dataset_id } = attributes;
+
+        let queryData = null;
+
+        if (query_id) {
+          // Get my queries, if they're stored
+          const { data } = await this.getQuery(query_id);
+          queryData = data;
+        } else {
+          // Otherwise, run the sql
+          const { sql } = attributes;
+          const { data } = await this.getData({ sql });
+          queryData = data;
+        }
+
+        let items = ''
+        if (typeof(queryData) === 'object') {
+          items = convertToCSV(queryData.data)
+        } else {
+          items = queryData
+        }
+
+        let datasetInfo = this.datasetsVisualizations.filter((dataset) => dataset.id == dataset_id)
+
+        const [{ attributes: { columns, slug, name: datasetName } }] = datasetInfo
+
+        // Append the visualization configuration
+        const visualization = { items, columns, slug, datasetName, config: spec, name, privacy_status, query_id, id, user_id, sql, dataset_id };
+
+        visualizations.push(visualization);
+      }
+
+      return visualizations;
+    },
+    removeAllIcons() {
+      /*Method to remove the config icon for all visualizations, we need to wait to load both lists when they are loaded, we select alls visualizations, and iterate over them with a loop to remove every icon.*/
+      this.$nextTick(() => {
+        let vizList = document.querySelectorAll("perspective-viewer");
+        for (let index = 0; index < vizList.length; index++) {
+          vizList[index].shadowRoot.querySelector("div#config_button").style.display = "none";
+        }
+      })
     }
   },
 };
